login 6/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pymongo==4.4.0
python-jose==3.3.0
passlib[bcrypt]>=1.7.4
pydantic==1.10.11
python-multipart==0.0.6
bcrypt==4.0.1
